Connectivity and visual fixes

+++ WebApi/Api/SignalR/AppHub_Connection.cs
@@ -2,57 +2,53 @@ using AutoMapper;
 using Microsoft.AspNetCore.SignalR;
 using System.Security.Claims;
 using WebApi.Api.DTO;
-using WebApi.Api.SignalR;
 using WebApi.Api.SignalR.Interfaces;
 using WebApi.Application.Services.Interfaces;
 using WebApi.Infrastructure.Repositories;
 
-namespace WebApi.SignalR;
-
-public partial class AppHub : Hub
+namespace WebApi.SignalR
 {
-    private readonly ILogger<AppHub> _logger;
-    private readonly IRoomRepository _roomRepository;
-    private readonly IChatRepository _chatRepository;
-    private readonly IPlaylistRepository _playlistRepository;
-    private readonly IUserRepository _userRepository;
-    private readonly IPlaylistService _playlistService;
-    private readonly IYouTubeAPIService _youtubeAPIService;
-    private readonly IJwtTokenService _jwtTokenService;
-    private readonly IVideoPlayerStateService _videoPlayerStateService;
-    private readonly IHubConnectionMapper _hubConnectionMapper;
-    private readonly IMapper _mapper;
-
-    public AppHub(
-        ILogger<AppHub> logger,
-        IPlaylistService playlistHandler,
-        IYouTubeAPIService youTubeAPIService,
-        IRoomRepository roomRepository,
-        IChatRepository chatRepository,
-        IPlaylistRepository playlistRepository,
-        IUserRepository userRepository,
-        IJwtTokenService tokenService,
-        IVideoPlayerStateService videoPlayerStateService,
-        IHubConnectionMapper hubConnectionMapper,
-        IMapper mapper
-    )
+    public partial class AppHub : Hub
     {
-        _logger = logger;
-        _playlistService = playlistHandler;
-        _youtubeAPIService = youTubeAPIService;
-        _roomRepository = roomRepository;
-        _chatRepository = chatRepository;
-        _playlistRepository = playlistRepository;
-        _userRepository = userRepository;
-        _jwtTokenService = tokenService;
-        _videoPlayerStateService = videoPlayerStateService;
-        _hubConnectionMapper = hubConnectionMapper;
-        _mapper = mapper;
-    }
+        private readonly ILogger<AppHub> _logger;
+        private readonly IRoomRepository _roomRepository;
+        private readonly IChatRepository _chatRepository;
+        private readonly IPlaylistRepository _playlistRepository;
+        private readonly IUserRepository _userRepository;
+        private readonly IVideoPlayerService _playlistService;
+        private readonly IYouTubeAPIService _youtubeAPIService;
+        private readonly IJwtTokenService _jwtTokenService;
+        private readonly IVideoPlayerStateService _videoPlayerStateService;
+        private readonly IHubConnectionMapper _hubConnectionMapper;
+        private readonly IMapper _mapper;
 
-    public override async Task OnConnectedAsync()
-    {
-        try
+        public AppHub(
+            ILogger<AppHub> logger,
+            IVideoPlayerService playlistHandler,
+            IYouTubeAPIService youTubeAPIService,
+            IRoomRepository roomRepository,
+            IChatRepository chatRepository,
+            IPlaylistRepository playlistRepository,
+            IUserRepository userRepository,
+            IJwtTokenService tokenService,
+            IVideoPlayerStateService videoPlayerStateService,
+            IHubConnectionMapper hubConnectionMapper,
+            IMapper mapper)
+        {
+            _logger = logger;
+            _playlistService = playlistHandler;
+            _youtubeAPIService = youTubeAPIService;
+            _roomRepository = roomRepository;
+            _chatRepository = chatRepository;
+            _playlistRepository = playlistRepository;
+            _userRepository = userRepository;
+            _jwtTokenService = tokenService;
+            _videoPlayerStateService = videoPlayerStateService;
+            _hubConnectionMapper = hubConnectionMapper;
+            _mapper = mapper;
+        }
+
+        public override async Task OnConnectedAsync()
         {
             var connectionId = Context.ConnectionId;
             var userId = GetUserId();
@@ -65,22 +61,22 @@ public partial class AppHub : Hub
                 return;
             }
 
-            _logger.LogInformation("User connected: {UserId}", userId);
+            _logger.LogInformation("User connected: {UserId} with ConnectionId: {ConnectionId}", userId, connectionId);
+
+            var previousConnections = _hubConnectionMapper.GetConnectionIdsByUserId(userId);
+
+            foreach (var prevConnectionId in previousConnections)
+            {
+                _hubConnectionMapper.CancelPendingDisconnection(userId, prevConnectionId);
+            }
 
             _hubConnectionMapper.AddUserConnection(userId, connectionId);
-
             await Groups.AddToGroupAsync(connectionId, roomHash);
+
             await base.OnConnectedAsync();
         }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "An error occurred in OnConnectedAsync");
-        }
-    }
 
-    public override async Task OnDisconnectedAsync(Exception exception)
-    {
-        try
+        public override async Task OnDisconnectedAsync(Exception exception)
         {
             var userId = GetUserId();
             var connectionId = Context.ConnectionId;
@@ -92,55 +88,69 @@ public partial class AppHub : Hub
                 return;
             }
 
-            await Task.Delay(5000);
+            var disconnectCancellationTokenSource = new CancellationTokenSource();
+            _hubConnectionMapper.TrackPendingDisconnection(userId, connectionId, disconnectCancellationTokenSource);
 
-            if (_hubConnectionMapper.GetConnectionIdByUserId(userId) == connectionId)
+            try
             {
-                var removedUser = await _userRepository.DeleteUserByConnectionIdAsync(roomHash, connectionId);
+                await Task.Delay(5000, disconnectCancellationTokenSource.Token);
 
-                _hubConnectionMapper.RemoveUserConnection(userId, connectionId);
-
-                await Groups.RemoveFromGroupAsync(connectionId, roomHash);
-
-                if (removedUser == null)
+                if (_hubConnectionMapper.IsConnectionPending(userId, connectionId))
                 {
-                    await base.OnDisconnectedAsync(exception);
-                    return;
+                    var remainingConnections = _hubConnectionMapper.GetConnectionIdsByUserId(userId);
+
+                    if (remainingConnections.Count >= 1)
+                    {
+                        _logger.LogInformation("Disconnecting user {UserId}, no reconnection within timeout.", userId);
+
+                        var removedUser = await _userRepository.DeleteUserByConnectionIdAsync(roomHash, connectionId);
+                        var room = await _roomRepository.GetRoomAsync(roomHash);
+
+                        _hubConnectionMapper.RemoveUserConnection(userId, connectionId);
+                        await Groups.RemoveFromGroupAsync(connectionId, roomHash);
+
+                        if (room != null && room.Users.Count == 0)
+                        {
+                            await _roomRepository.DeleteRoomAsync(roomHash);
+                            _logger.LogInformation("Room {RoomHash} deleted as no users remain.", roomHash);
+                        }
+
+                        if (removedUser != null)
+                        {
+                            var userDTO = _mapper.Map<UserDTO>(removedUser);
+                            await Clients.Group(roomHash).SendAsync(HubMessages.OnLeaveRoom, userDTO);
+
+                            _logger.LogInformation("User {UserId} disconnected and removed from room {RoomHash}.", userId, roomHash);
+                        }
+                    }
                 }
-
-                var userDTO = _mapper.Map<UserDTO>(removedUser);
-
-                await Clients.Group(roomHash).SendAsync(HubMessages.OnLeaveRoom, userDTO);
-
-                var room = await _roomRepository.GetRoomAsync(roomHash);
-
-                if (room.Users.Count == 0)
+                else
                 {
-                    await _roomRepository.DeleteRoomAsync(roomHash);
+                    _logger.LogInformation("Reconnection occurred for user {UserId}, canceling disconnection.", userId);
                 }
-
-                _logger.LogInformation("Hub: User Disconnected: {ConnectionId}", connectionId);
             }
-            else
+            catch (TaskCanceledException)
             {
-                _logger.LogInformation("Hub: User {UserId} has another active connection in the same room", userId);
+                _logger.LogInformation("Disconnection canceled, user {UserId} reconnected.", userId);
             }
-
-            await base.OnDisconnectedAsync(exception);
+            finally
+            {
+                _hubConnectionMapper.ClearPendingDisconnection(userId, connectionId);
+                await base.OnDisconnectedAsync(exception);
+            }
         }
-        catch (Exception ex)
+
+
+
+
+        private string GetUserId()
         {
-            _logger.LogError(ex, "An error occurred in OnDisconnectedAsync");
+            return Context.User?.FindFirst(ClaimTypes.NameIdentifier)?.Value;
         }
-    }
 
-    private string GetUserId()
-    {
-        return Context.User?.FindFirst(ClaimTypes.NameIdentifier)?.Value;
-    }
-
-    private string GetRoomHash()
-    {
-        return Context.User?.FindFirst(ClaimTypes.Hash)?.Value;
+        private string GetRoomHash()
+        {
+            return Context.User?.FindFirst(ClaimTypes.Hash)?.Value;
+        }
     }
 }
